fix: add workaround to handle inline quotes in the richtext field of news stories
NOTE: currently the children of the inline quote doesn't render correctly when
using the blockquote component in the nodeResolver of
storyblok-rich-text-renderer. We bump into a few issues around rendering
and correct typing. I've hardcoded the html and imported
styles so we can render the correctly styled quote

TODO: Update the blockquote component or richtext types to handle
passing through richtext quote children to the blockquote component