NXT-426 Fixed fetchAndMapContentParts function
fetchAndMapContentParts function was recently added for the feed update - contentPartList2.

Function was returning an array regardless of whether it had found the correct type using the filter. Function now returns undefined and will return an object as well as an array.

The types for contentParts has also been updated to allow an object as well as an array for when there are multiple parts.